requirements.txt
flask==3.0.3
flask-dance==7.1.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-talisman==1.1.0
psycopg[binary]==3.2.13
gunicorn==22.0.0
python-dotenv==1.0.1
sendgrid==6.11.0
